Slightly change data structures
I tried yet again to implement `get_line(..)`. It failed horribly. But
in the struggle, I learned a few things, so I kept some changes. The
main thing is that fragments now use a vector instead of a hashmap to
keep track of the line byte offsets, and that the byte offsets are
relative to the fragment's own offset values.