Render something to the screen
I've been thinking about different ways to model the rendering logic.
This is gonna affect a ton of stuff and I really wanna get it right. But
I also wanna ship code. None of this code is useful, but it makes me
feel like I actually accomplished something, so there's that.